requirements.txt
Django==4.2.2
django-rest-swagger==2.2.0
djangorestframework==3.14.0
PyYAML==6.0
